cli/requirements.txt
pyserial
ansicolor
